+++ test/integration/069_build_test/test_build.py
@@ -2,14 +2,10 @@ from test.integration.base import DBTIntegrationTest, use_profile
 import yaml
 
 
-class TestBuild(DBTIntegrationTest):
+class TestBuildBase(DBTIntegrationTest):
     @property
     def schema(self):
         return "build_test_069"
-
-    @property
-    def models(self):
-        return "models"
 
     @property
     def project_config(self):
@@ -31,24 +27,55 @@ class TestBuild(DBTIntegrationTest):
 
         return self.run_dbt(args, expect_pass=expect_pass)
 
+
+class TestPassingBuild(TestBuildBase):
+    @property
+    def models(self):
+        return "models"
+
     @use_profile("postgres")
     def test__postgres_build_happy_path(self):
         self.build()
 
 
-class TestFailingBuild(TestBuild):
-    @property
-    def schema(self):
-        return "build_test_069"
-
+class TestFailingBuild(TestBuildBase):
     @property
     def models(self):
         return "models-failing"
-        
+
     @use_profile("postgres")
     def test__postgres_build_happy_path(self):
         results = self.build(expect_pass=False)
-        self.assertEqual(len(results), 12)
+        self.assertEqual(len(results), 13)
         actual = [r.status for r in results]
-        expected = ['error']*1 + ['skipped']*4 + ['pass']*2 + ['success']*5
+        expected = ['error']*1 + ['skipped']*5 + ['pass']*2 + ['success']*5
+        self.assertEqual(sorted(actual), sorted(expected))
+
+
+class TestFailingTestsBuild(TestBuildBase):
+    @property
+    def models(self):
+        return "tests-failing"
+
+    @use_profile("postgres")
+    def test__postgres_failing_test_skips_downstream(self):
+        results = self.build(expect_pass=False)
+        self.assertEqual(len(results), 13)
+        actual = [str(r.status) for r in results]
+        expected = ['fail'] + ['skipped']*6 + ['pass']*2 + ['success']*4
+        self.assertEqual(sorted(actual), sorted(expected))
+
+
+class TestCircularRelationshipTestsBuild(TestBuildBase):
+    @property
+    def models(self):
+        return "models-circular-relationship"
+
+    @use_profile("postgres")
+    def test__postgres_circular_relationship_test_success(self):
+        """ Ensure that tests that refer to each other's model don't create
+        a circular dependency. """
+        results = self.build()
+        actual = [r.status for r in results]
+        expected = ['success']*7 + ['pass']*2
         self.assertEqual(sorted(actual), sorted(expected))